requirements.txt
discord.py==2.5.2
icecream==2.1.5